fix folder paths
