--- src/components/cardInfo.module.css
@@ -1,155 +1,0 @@
-.cardInfo{
-	display: flex;
-	gap: 40px;
-	height: 1100px;
-	padding: 40px;
-	width: 1340px;
-	margin: 0 auto;
-	color: #d1c9c9;
-}
-.video{
-	width: 100%;
-	max-height: 370px;
-	background-color: rgba(125, 111, 111, 0.10);
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	flex: 1;
-}
-.info{
-	width: 540px;
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-}
-.title{
-	height: 130px;
-	margin-top: 22px;
-}
-.infoBlock{
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	margin: 24px 0px;
-}
-.peas{
-	display: flex;
-	align-items: center;
-	gap: 14px;
-}
-.fullhd{
-	background-color: rgba(125, 111, 111, 0.20);
-	padding: 2px;
-	border-radius: 2px;
-}
-.ratingAndActors{
-	display: flex;
-	align-items: center;
-	justify-content: space-between;
-	width: 440px;
-	cursor: pointer;
-}
-.squers{
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	text-align: center;
-	width: 60px;
-	gap: 8px;
-	transition: color 0.3s;
-}
-.squers:hover{
-	color: white;
-}
-.squers:hover .squer{
-	color: white;
-	box-shadow: 0px 0px 50px rgba(255, 255, 255, 0.16);
-}
-.squer{
-	background-color: gray;
-	width: 50px;
-	height: 50px;
-	border-radius: 8px;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	font-weight: bold;
-	font-size: 20px;
-	transition: box-shadow 0.3s;
-}
-.actors{
-	font-size: 14px;
-}
-.actions{
-	display: flex;
-	justify-content: space-between;
-	align-items: center;
-	gap: 8px;
-	width: 100%;
-	margin: 26px 0px;
-}
-.btn{
-	background-color: rgba(125, 111, 111, 0.20);
-	border: none;
-	border-radius: 8px;
-	padding: 10px;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	gap: 8px;
-	font-size: 16px;
-	transition: background-color 0.3s;
-}
-.btn:hover{
-	background-color: rgba(125, 111, 111, 0.28);
-}
-.triler{
-	flex: 1;
-}
-.ratingBlock{
-	display: flex;
-	justify-content: space-between;
-	background-color: rgba(125, 111, 111, 0.20);
-	width: 100%;
-	border-radius: 8px;
-	padding: 10px;
-	cursor: pointer;
-	margin: 30px;
-}
-.ratingCube{
-	background-color: #73A32F;
-	width: 66px;
-	height: 66px;
-	border-radius: 8px;
-	font-size: 22px;
-	font-weight: bold;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	color: white;
-}
-.ratingDesc{
-	flex: 1;
-	margin-left: 18px;
-	display: flex;
-	flex-direction: column;
-	justify-content: center;
-}
-.ratingTitle{
-	color: white;
-	font-weight: bold;
-}
-.ratingBtnCont{
-	display: flex;
-	align-items: center;
-	justify-content: center;
-}
-.ratingBtn{
-	background-color: transparent;
-	border: 1px solid gray;
-	border-radius: 6px;
-	padding: 4px;
-	cursor: pointer;
-	font-weight: bold;
-	margin-right: 10px;
-}
