+++ src/pages/HomePage/index.tsx
@@ -1,4 +1,4 @@
-import { Cards } from '@/components/Cards';
+import { Cards } from '@/components/Cards/Cards';
 
 export function HomePage() {
   return <Cards />;
