+++ src/pages/Movies/index.tsx
@@ -1,6 +1,6 @@
 import { Link } from 'react-router-dom';
 import { Search } from '@/components/Search';
-import { Cards } from '@/components/Cards';
+import { Cards } from '@/components/Cards/Cards';
 
 export function Movies() {
   return <h1>Фильмы смотреть онлайн</h1>;
