+++ src/pages/layout/index.tsx
@@ -2,8 +2,7 @@ import styles from './styles.module.css';
 import { Header, Footer } from '@/widgets';
 import { Outlet } from 'react-router-dom';
 import { Home } from '../home';
-import { Cards } from '../../components/Cards';
-import CardInfo from '../../components/CardInfo';
+import { Cards } from '@/components/Cards/Cards';
 
 export const Layout = () => {
   return (
@@ -12,7 +11,7 @@ export const Layout = () => {
       <main className={styles.main}>
         <div className={styles.main__content}>
           <Outlet />
-          <CardInfo/>
+					<Cards/>
         </div>
       </main>
       <Footer />
